Started re-writing for new modular system. Hopefully this time I'll actually finish it :)
