+++ ColorChord.NET/ColorChord.NET.csproj
@@ -8,9 +8,10 @@
     <OutputType>WinExe</OutputType>
     <RootNamespace>ColorChord.NET</RootNamespace>
     <AssemblyName>ColorChord.NET</AssemblyName>
-    <TargetFrameworkVersion>v4.6.1</TargetFrameworkVersion>
+    <TargetFrameworkVersion>v4.6</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <AutoGenerateBindingRedirects>true</AutoGenerateBindingRedirects>
+    <TargetFrameworkProfile />
   </PropertyGroup>
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
     <PlatformTarget>AnyCPU</PlatformTarget>
@@ -36,7 +37,7 @@
     <ApplicationIcon>Colochord2-Logo.ico</ApplicationIcon>
   </PropertyGroup>
   <PropertyGroup>
-    <StartupObject />
+    <StartupObject>ColorChord.NET.ColorChord</StartupObject>
   </PropertyGroup>
   <ItemGroup>
     <Reference Include="System" />
@@ -54,17 +55,25 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="AudioTools.cs" />
-    <Compile Include="LinearOutput.cs" />
+    <Compile Include="ColorChord.cs" />
+    <Compile Include="NoteFinder.cs" />
+    <Compile Include="OLD\AudioTools.cs" />
+    <Compile Include="OLD\LinearOutput.cs" />
     <Compile Include="MainForm.cs">
       <SubType>Form</SubType>
     </Compile>
     <Compile Include="MainForm.Designer.cs">
       <DependentUpon>MainForm.cs</DependentUpon>
     </Compile>
-    <Compile Include="NoteFinder.cs" />
-    <Compile Include="Program.cs" />
+    <Compile Include="OLD\NoteFinder.cs" />
+    <Compile Include="OLD\Program.cs" />
+    <Compile Include="Outputs\IOutput.cs" />
+    <Compile Include="Outputs\PacketUDP.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Sources\IAudioSource.cs" />
+    <Compile Include="Sources\WASAPILoopback.cs" />
+    <Compile Include="Visualizers\IVisualizer.cs" />
+    <Compile Include="Visualizers\Linear.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
