+++ ColorChord.NET/ColorChord.cs
@@ -1,0 +1,18 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace ColorChord.NET
+{
+    public class ColorChord
+    {
+
+        public static void Main(string[] args)
+        {
+
+        }
+
+    }
+}
