+++ ColorChord.NET/OLD/AudioTools.cs
