+++ ColorChord.NET/OLD/LinearOutput.cs
@@ -269,7 +269,7 @@ namespace ColorChord.NET
         }
 
         private static UdpClient Sender = new UdpClient();
-        private static IPEndPoint Destination = new IPEndPoint(IPAddress.Parse("192.168.0.102"), 7777);
+        private static IPEndPoint Destination = new IPEndPoint(IPAddress.Parse("192.168.0.60"), 7777);
 
         public static void Send()
         {
