+++ ColorChord.NET/Visualizers/Linear.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace ColorChord.NET.Visualizers
+{
+    public class Linear : IVisualizer
+    {
+
+
+    }
+}
